+++ lib/src/widget/snacky_configurator_widget.dart
@@ -20,11 +20,14 @@ class SnackyConfiguratorWidget extends StatefulWidget {
   });
 
   @override
-  State<SnackyConfiguratorWidget> createState() => _SnackyConfiguratorWidgetState();
+  State<SnackyConfiguratorWidget> createState() =>
+      _SnackyConfiguratorWidgetState();
 }
 
-class _SnackyConfiguratorWidgetState extends State<SnackyConfiguratorWidget> implements SnackyListener {
-  SnackyController get snackyController => widget.snackyController ?? SnackyController.instance;
+class _SnackyConfiguratorWidgetState extends State<SnackyConfiguratorWidget>
+    implements SnackyListener {
+  SnackyController get snackyController =>
+      widget.snackyController ?? SnackyController.instance;
 
   @override
   void initState() {
@@ -72,7 +75,8 @@ class _SnackyConfiguratorWidgetState extends State<SnackyConfiguratorWidget> imp
 
     context.visitChildElements(visitor);
 
-    assert(navigator != null, '''It looks like you are not using Navigator in your app.
+    assert(navigator != null,
+        '''It looks like you are not using Navigator in your app.
          Do you wrapped you app widget like this?
          SnackyConfiguratorWidget(
            app: MaterialApp(